requirements.txt
pymysql
requests
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask_JWT_Extended
Flask_SocketIO
Flask_Logging
Flask-Login 
python-dotenv
# flask-seeder
sqlalchemy-seeder
openpyxl
reportlab
python-bidi
arabic-reshaper
cryptography
jdatetime
dateutils 